to_mqtt/lfs_mqtt/requirements.txt
paho-mqtt==2.1.0
PyYAML==6.0.1
watchdog==4.0.1
